Move ServerResponse to core HTTP
