+++ lib/mod.js
@@ -5,6 +5,7 @@ export { default as JSModuleConfigStore } from './config-stores/js-module-config
 export { default as ApplicationContext } from './context/application-context.js';
 export { default as RequestContext } from './context/request-context.js';
 export { default as HttpRouter } from './http-router/http-router.js';
+export { default as ServerResponse } from './http/server-response.js';
 export { default as JSModuleHttpRoutesStore } from './http-routes-stores/js-module-http-routes-store.js';
 export * as NodeLocalHyperview from './hyperview/node-local-store/plugin.js';
 export { default as BaseLogger } from './logger/base-logger.js';
