+++ lib/node-http-server/node-server.js
@@ -3,7 +3,7 @@ import { EventEmitter } from 'node:events';
 import { AssertionError } from '../errors.js';
 import { isNumberNotNaN, isString, isUndefined, isFunction } from '../assertions.js';
 import ServerRequest from './server-request.js';
-import ServerResponse from './server-response.js';
+import ServerResponse from '../http/server-response.js';
 
 /**
  * Payload for NodeServer 'debug' events (request-received, response-sent).
